rename `component` mod to `fn_component`; cargo fmt
