+++ crates/kobold_macros/src/fn_component.rs
